+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # chat21-web-widget ver 3.0
 
 ### 4.0.47
-- new: added Policy Privacy
+- new: added privacy policy
 
 ### 4.0.46A
 - bug-fix: right alignment of the buttons